Semi-major changes to code- solar tracker has its own tab
SolarTracker function is now in a separate tab and a variable returning
function, which the main loop compares whether it went through its
tracking routine or not. If tracking was done, that is the delay
between xively calls (about 18 seconds). If the panel did not track due
to traveling speed, a delay of 15 seconds from the main loop is used.
Serial output now includes status of relay outputs.